proiect-retele-2023-imr/requirements.txt
asttokens==2.2.1
backcall==0.2.0
blinker==1.6.2
branca==0.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cryptography==41.0.1
cycler==0.11.0
decorator==5.1.1
dnslib==0.9.23
dnspython==2.3.0
docker==6.1.3
executing==1.2.0
Flask==2.3.2
folium==0.14.0
fonttools==4.39.4
future==0.18.3
geocoder==1.38.1
geographiclib==2.0
geopy==2.3.0
idna==3.4
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
numpy==1.24.3
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.38
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
pyparsing==3.1.0b2
python-dateutil==2.8.2
pywin32==306
PyX==0.16
ratelim==0.1.6
requests==2.31.0
scapy==2.5.0
six==1.16.0
stack-data==0.6.2
traitlets==5.9.0
urllib3==2.0.2
wcwidth==0.2.6
websocket-client==1.5.2
Werkzeug==2.3.6
netfilterqueue==0.8.1